Move class to separate files
