+++ samples/web/send-from-nancy/Core_6/WebApplication/Bootstraper.cs
@@ -1,36 +1,9 @@
 ﻿using Nancy;
 using Nancy.TinyIoc;
 using NServiceBus;
-using System.Threading.Tasks;
 
 namespace WebApplication
 {
-    public class SendMessageModule : NancyModule
-    {
-        private readonly IMessageSession messageSession;
-
-        public SendMessageModule(IMessageSession messageSession) : base("/sendMessage")
-        {
-            this.messageSession = messageSession;
-            
-            this.Get["/", true] = async (r, c) => 
-            {
-                var message = new MyMessage();
-                await messageSession.Send(message)
-                    .ConfigureAwait(false);
-                return "Message sent to endpoint";
-            };
-        }
-    }
-
-    public class RootModule : NancyModule
-    {
-        public RootModule() : base()
-        {
-            this.Get["/"] = r => this.Response.AsRedirect("/sendMessage");
-        }
-    }
-
     public class Bootstraper : DefaultNancyBootstrapper
     {
         protected override void ConfigureApplicationContainer(TinyIoCContainer container)
@@ -61,22 +34,11 @@ namespace WebApplication
 
             #endregion
 
+            #region ServiceRegistration
+
             container.Register<IMessageSession>(endpointInstance);
-        }
-    }
 
-    public interface IDbContext
-    {
-        string GetData();
-    }
-
-    public class DbContext : IDbContext
-    {
-        private int count = 0;
-
-        public string GetData()
-        {
-            return (++count).ToString();
+            #endregion
         }
     }
 }
