+++ samples/web/send-from-nancy/Core_6/WebApplication/Modules/RootModule.cs
@@ -1,0 +1,12 @@
+﻿using Nancy;
+
+namespace WebApplication.Modules
+{
+    public class RootModule : NancyModule
+    {
+        public RootModule() : base()
+        {
+            this.Get["/"] = r => this.Response.AsRedirect("/sendMessage");
+        }
+    }
+}
