+++ samples/web/send-from-nancy/Core_6/WebApplication/Modules/SendMessageModule.cs
@@ -1,0 +1,23 @@
+﻿using Nancy;
+using NServiceBus;
+
+namespace WebApplication.Modules
+{
+    public class SendMessageModule : NancyModule
+    {
+        private readonly IMessageSession messageSession;
+
+        public SendMessageModule(IMessageSession messageSession) : base("/sendMessage")
+        {
+            this.messageSession = messageSession;
+            
+            this.Get["/", true] = async (r, c) => 
+            {
+                var message = new MyMessage();
+                await messageSession.Send(message)
+                    .ConfigureAwait(false);
+                return "Message sent to endpoint";
+            };
+        }
+    }
+}
