+++ samples/web/send-from-nancy/Core_6/WebApplication/WebApplication.csproj
@@ -45,8 +45,10 @@
     <PackageReference Include="NServiceBus" Version="6.*" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="CommentModule.cs" />
+    <Compile Include="Bootstraper.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Modules\RootModule.cs" />
+    <Compile Include="Modules\SendMessageModule.cs" />
     <Content Include="Web.config" />
   </ItemGroup>
   <ItemGroup>
